chore(samples): Reorganize the samples section of the project structure
After adjusting the project structure of the samples for the first time, each individual sample under the samples was modified to fit the new sample publishing method.
