+++ Assets/Plugins/PsychoUnity/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 79982c8388e850245ab58843655e23a2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
