+++ Assets/Samples.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 302c01eec94be1d4d80f7b67ae332d2a
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
