--- Assets/Samples/Events/Action2.cs
@@ -1,18 +1,0 @@
-using PsychoUnity.Manager;
-using UnityEngine;
-
-namespace Samples.Events
-{
-    public class Action2 : MonoBehaviour
-    {
-        void Start()
-        {
-            EventManager.Instance.AddEventListener("UnderAttack", BloodState);
-        }
-
-        private void BloodState()
-        {
-            print("Blooding...");
-        }
-    }
-}
