+++ Assets/Samples/Events/Scenes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f55b131f576d50041a0d717ba7a469bb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
