+++ Assets/Samples/Events/Scripts.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: d675380c58aa5c84ea46b28591161409
+guid: e168d7703b2de7c4d9f411da1799610c
 folderAsset: yes
 DefaultImporter:
   externalObjects: {}
