--- Assets/Samples/HighResolutionTimer/SceneLoader.cs
@@ -1,23 +1,0 @@
-using System;
-using UnityEngine;
-using UnityEngine.SceneManagement;
-
-namespace Samples.HighResolutionTimer
-{
-    public class SceneLoader : MonoBehaviour
-    {
-        public GameObject prefix;
-        private void Awake()
-        {
-            DontDestroyOnLoad(prefix);
-        }
-
-        private void Update()
-        {
-            if (Input.GetKeyDown(KeyCode.Space))
-            {
-                SceneManager.LoadScene("Samples/HighResolutionTimer/SubScene");
-            }
-        }
-    }
-}
