+++ Assets/Samples/Recorder/Data.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: db11469d09909254ba5d8db93945ff36
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
