+++ Assets/Samples/Recorder/Data/ExampleDataStructure.csv.meta
