+++ Assets/Samples/Recorder/Scenes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: d0ecf44e970ff2848a3bb42dc92bbc29
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
