+++ Assets/Samples/Recorder/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f30589afb01a5434dae7c4b2f0a2b5bb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
