+++ Assets/Samples/TCPCom/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6f5990114fd7b6549a09aae4a394aa1b
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
