+++ Assets/Samples/Timer/Scenes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 10a3b5130366338498a7c935e3ecf9df
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
