+++ Assets/Samples/Timer/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8215a8ba884f85b42aa30bb20f7a4470
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
