+++ Assets/Samples/TimerHighResolution/Scenes.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 6cad191303fd68a49a00c0439a3d0047
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
