+++ Assets/Samples/TimerHighResolution/Scenes/Main.unity
