+++ Assets/Samples/TimerHighResolution/Scripts.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 0ed3c3cfd7cfebc41a67f0240bf528e9
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
