+++ Assets/Samples/TimerHighResolution/Scripts/Logic.cs.meta
