+++ Assets/Samples/TimerHighResolution/Scripts/SceneLoader.cs
@@ -1,0 +1,22 @@
+using UnityEngine;
+using UnityEngine.SceneManagement;
+
+namespace Samples.TimerHighResolution.Scripts
+{
+    public class SceneLoader : MonoBehaviour
+    {
+        public GameObject prefix;
+        private void Awake()
+        {
+            DontDestroyOnLoad(prefix);
+        }
+
+        private void Update()
+        {
+            if (Input.GetKeyDown(KeyCode.Space))
+            {
+                SceneManager.LoadScene("Samples/HighResolutionTimer/SubScene");
+            }
+        }
+    }
+}
