+++ Assets/Samples/TimerHighResolution/Scripts/SceneLoader.cs.meta
