+++ Assets/Samples/TimerHighResolution/Scripts/Timer.cs.meta
